+++ packages/cli-data/src/modules/party/commands/index.ts
@@ -1,0 +1,11 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+export * from './create';
+export * from './info';
+export * from './invite';
+export * from './join';
+export * from './list';
+export * from './members';
+export * from './open';
